repr: impl From<Datum> for serde_json::Value
